fix(server): move @before_compile hook out of ClickHouseRepo
`@before_compile __MODULE__` doesn't work when the macro is defined
in the same module being compiled — Elixir can't resolve the
`__before_compile__/1` callback because the module isn't fully loaded
when the callback fires. CI hit this with `function
Tuist.ClickHouseRepo.__before_compile__/1 is undefined`.

Move the query / query! retry hook into a sibling
`Tuist.ClickHouseRepo.QueryRetry` module and reference it as
`@before_compile`. The sibling module is fully compiled by the time
the callback fires, so the macro is resolvable.
